+++ flexmeasures/ui/views/assets/forms.py
@@ -37,6 +37,10 @@ class AssetForm(FlaskForm):
         places=None,
         render_kw={"placeholder": "--Click the map or enter a longitude--"},
     )
+    parent_asset_id = IntegerField(
+        "Parent Asset Id",
+        validators=[optional()],
+    )
     attributes = StringField(
         "Other attributes (JSON)",
         default="{}",
@@ -122,7 +126,6 @@ class NewAssetForm(AssetForm):
         "Asset type", coerce=int, validators=[DataRequired()]
     )
     account_id = SelectField("Account", coerce=int, validators=[optional()])
-    parent_asset_id = IntegerField("Parent Asset Id", validators=[optional()])
 
     def set_account(self) -> tuple[Account | None, str | None]:
         """Set an account for the to-be-created asset.